delete unnessary import, delete not used method
