+++ myplugin/content/admin_instructions/panel.py
@@ -1,4 +1,3 @@
-from django.utils.translation import gettext_lazy as _
 import horizon
 
 class AdminInstructionsPanel(horizon.Panel):
