+++ myplugin/content/admin_instructions/views.py
@@ -1,57 +1,7 @@
-from http.client import responses
-
-from django.http import JsonResponse
-from django.utils.decorators import method_decorator
 from django.views import generic
 from django.utils.translation import gettext_lazy as _
-
-from myplugin.content.api_endpoints import API_ENDPOINTS
-
-import requests
-import logging
+from myplugin.content.admin_instructions.utils import get_user_details, get_roles
 import sys
-
-from django.views.decorators.csrf import csrf_exempt
-
-def get_token_id(request):
-    """
-    Retrieves the token ID from the request object.
-    """
-    return getattr(getattr(request, "user", None), "token", None) and request.user.token.id
-
-def get_user_details(request, user_id):
-    """
-    Fetches detailed user information for a given user_id using the external API.
-    """
-    token_id = get_token_id(request)
-    headers = {"X-Auth-Token": token_id}
-    url = f"{API_ENDPOINTS['user_list']}{user_id}"
-
-    try:
-        response = requests.get(url, headers=headers, timeout=10)
-        response.raise_for_status()
-        return response.json()
-    except requests.RequestException as e:
-        logging.error("Failed to fetch user details for user_id %s: %s", user_id, e)
-        return {}
-
-def get_roles(request):
-    """
-    Fetches app templates from the external API using a provided token ID.
-    """
-    token_id = get_token_id(request)
-    headers = {"X-Auth-Token": token_id}
-
-    try:
-        response = requests.get(API_ENDPOINTS['roles_list'],
-                                headers=headers, timeout=10)
-        response.raise_for_status()
-        return response.json()
-    except requests.RequestException as e:
-        logging.error("Failed to fetch roles: %s", e)
-        return []
-
-
 
 class IndexView(generic.TemplateView):
     """
