+++ myplugin/content/example/panel.py
@@ -1,4 +1,3 @@
-from django.utils.translation import gettext_lazy as _
 import horizon
 
 class ExamplePanel(horizon.Panel):
