+++ myplugin/content/example/views.py
@@ -1,12 +1,5 @@
-from django.http import JsonResponse
-from django.utils.decorators import method_decorator
 from django.views import generic
 from django.utils.translation import gettext_lazy as _
-
-import requests
-from django.views.decorators.csrf import csrf_exempt
-
-#THIS CODE IS NOT USED AT THE MOMENT
 
 class IndexView(generic.TemplateView):
     """
