+++ myplugin/content/tutorial/views.py
@@ -1,15 +1,5 @@
-from http.client import responses
-
-from django.http import JsonResponse
-from django.utils.decorators import method_decorator
 from django.views import generic
 from django.utils.translation import gettext_lazy as _
-
-import requests
-from django.views.decorators.csrf import csrf_exempt
-
-
-
 
 class IndexView(generic.TemplateView):
     """
